Reorder loops to increase speed
Re-running the role takes a quite long time as all client keys, certificate and
config files are re-generated. If we rather assume that the triple of client
key, cert and config exists when the key exists, the steps can be skipped for
existing configurations. The removal of all files is assured by the revocation
step.